Replaced input system, replaced all malloc and frees
Also gave several structs default initializers, and updated the sizes and signs of a lot of variables.

When designing the new input system, at first I used >> overloads. But it was too hard to keep track of what statements were advancing 2 bytes, and which were advancing 4 bytes. So I switched to a more explicit method.